Do not save z_table to yaml config
When resuming from a checkpoint the z_table will be loaded
to the config namespace.
Before this commit the z_table was saved in the yaml config
as AtomicTable, which can't be loaded if you try to continue from that
config file.
With this commit the z_table is saved as None (null) in the config.