Unified CLI.
This commit creates a single command interface "e3mat" for the package.
Different types of functionalities will be subcommands of the CLI. E.g.
the previous `matrixmace` is (for now) `e3mat models mace main`.

A dedicated directory has been created for the CLI (which is implemented
using typer) and packages are imported lazily. This is so that the CLI
can be as fast as possible (Some parts of e3nn_matrix take significant
time to load).

The exact architecture and naming in the CLI are still subject to change,
this is just a first draft.