Major refactor to reuse functionalities for the (enhanced) CLI.
Most things have been moved around to facilitate modularity
and reuse of functionalities, so that the code is not restricted
only to the use cases that we have thought about.

As a result, this commit is able to introduce the implementation of a
production ready server using fastapi. (the technical part is production
ready, not the API endpoints, which will still probably change a lot :) ).

The CLI then incorporates commands to launch this server and easily request
the most common things to it.