Lightning CLI: Checkpoint file now determines the defaults for arguments.
Any argument that is not passed to the CLI will be defaulted to whatever
was stored in the ckpt file. Additionally, one can now override the
checkpoint arguments, as they are simply used as defaults.

This is FINALLY a huge improvement on the usability of the CLI. Previously
you'd have to either pass all the arguments or stick with exactly the arguments
in the checkpoint file, which was problematic because of arguments like `root_dir`.