Added symmetrizer argument to MatrixBlock.
Edge readouts now should have the `symm_transpose` argument if they support
doing the operation in a way that f(x, y) == f(y, x).T . If they do not support
it, MatrixBlock does the symmetrization by computing both directions.

This symmetrization is only needed for edges between atoms of the same type, because
for the other cases we are certain that the operation is always done in the same
direction (nodes are sorted by type).