Fixed HUGE BUG on node features for edge blocks
In a system with edges `[[0, 1, ...], [1, 0, ...]]`, `i_edges` (forward edge direction) is `[0, ...]` and `j_edges` (backward edge direction) is `[1, ...]`.
Imagine atom 0 is type 0 and atom 1 is type 1. Then `type_edge_index` is `[[0, 1, ...], [1, 0, ...]]`. To get the pair of nodes of the forward edge, I would have to get the first item from each row, that is [0,1]. Instead, I was getting the first index from the first row and the second index from the second row. That is [0,0]. So effectively it only considered one of the nodes!!! 

I'm not sure it affects the results though as I think the models we are using don't use node features for the edge blocks.