Make intialization of E3nnGraph2Mat much faster
The initialization of tensor products in e3nn was too slow,
and this was making E3nnGraph2Mat unreasonably slow. This PR
fixes it by introducing a custom ReducedTensorProducts class
that precomputes and reuses rtps. There is even the possibility
to precompute the rtps once and store them on disk by doing:

graph2mat utils precompute-rtps --max-l 4